Add files via upload
This is the updated bios I used to split the ram as needed. More testing on this bios is required to see what options actually work and what options do nothing at all. Use at you own risk and make a copy of your original bios for backup. I am not responsible if you brick your board nor am I obliged to help you in any way to fix it.